Code refactoring, restructuring and optimization. Implemented the 'acb_content_block_title' filter. Protected the core plugin class functions (public => private). PHPDoc fixes. Bumped the version number to 0.2.0.
